Fix minimap not rendering tiles on edges

+++ src/render/utils/put_fill_rect.c
@@ -6,7 +6,7 @@
 /*   By: rreimann <rreimann@student.42heilbronn.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 15:39:09 by rreimann          #+#    #+#             */
-/*   Updated: 2025/05/01 23:44:38 by rreimann         ###   ########.fr       */
+/*   Updated: 2025/05/05 18:52:35 by rreimann         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 // This is filling a rect assuming no rotation
 void	put_fill_rect(mlx_image_t *img, t_rect *rect, uint32_t color)
 {
-	size_t	x;
-	size_t	y;
+	int	x;
+	int	y;
 
 	x = rect->vertices[0].x;
 	while (x < rect->vertices[1].x)
